Support indexing and searching function/method definitions from ELF binaries.
This commit addresses the weakness of "gj -d" which finds the declaration
or definitions based on heuristic patterns. The search is slow and there
are too many results when searching a general term such as "Save".

If the goal is to find the definitions, we can do it effciently and accurately
by using the debug info in ELF binaries,

This commit includes:
* Add "gj_index.py PROGRAM" to index ELF binaries' info about symbols in the text secion.
* Add "gj -D" to find out function/method's definition based on the debug info in ELF binaries.

Both the search and inex time are significantly faster (<0.1s) for a large code base.