Fix bug that some symbols and file line are mismatched due to objdump's bug.
The root cause is that objdump truncate the output if the symbol or directory is too long.
The maximum length are 35 and 76 hard coded in binutils/dwarf.c.
In theory, "--wide" should relax such limit. However, objdump doesn't treat the flag right.
On the other hand, readelf can do the same thing and "readelf --wide" works.
Use readelf to bypass the issue.