Show the item index continuously when folding is enabled.
Previously the item indexes are the same no matter folding is enabled or not
(folding is enabled via "." in the interactive mode). This commit always list
items as 1, 2, 3, ..., N where N is the total number of items printed to the console.

Also fix the bug that the selected items are not marked when selecting multiple items.